AI/requirements.txt
python-dotenv
confluent-kafka
boto3
openai-whisper
